@@ -7,19 +7,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^29.5.4",
-    "@types/node": "^20.5.9",
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
-    "@typescript-eslint/parser": "^6.6.0",
     "body-scroll-lock": "^3.1.3",
-    "framer-motion": "^2.3.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "framer-motion": "^10.16.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-intersection-observer": "^0.1.0",
-    "react-intersection-observer": "^8.27.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.1"
+    "react-intersection-observer": "^9.5.2",
+    "react-router": "^6.16.0",
+    "react-router-dom": "^6.16.0"
+    
   },
   "overrides": {
     "react-scripts": {
@@ -29,7 +25,7 @@
   "scripts": {
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -50,8 +46,14 @@
     ]
   },
   "devDependencies": {
+    "@types/jest": "^29.5.5",
+    "@types/node": "^20.8.4",
+    "@types/react": "^18.2.27",
+    "@types/react-dom": "^18.2.12",
+    "@types/react-router-dom": "^5.3.3",
     "gh-pages": "^3.1.0",
     "sass": "^1.50.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "react-scripts": "^5.0.1"
   }
 }
