@@ -64,9 +64,9 @@ function Banner() {
             viewBox="0 0 24 24"
             fill="none"
             stroke="currentColor"
-            stroke-width="2"
-            stroke-linecap="round"
-            stroke-linejoin="round"
+            strokeWidth={2}
+            strokeLinecap="round"
+            strokeLinejoin="round"
             className="feather feather-chevron-down"
             animate={{ y: 10 }}
             transition={{ yoyo: Infinity, duration: 1 }}
